♻️ Refactor Notion API code (#2)
* 🚚 Change notion.ts to notion/index.ts

* ♻️ Create formatDatabaseProperty

* 🚚 Sort notion API calls

This commit moves around a lot of code, but there is no change in functionality.

* 🚚 Move notion types to associated files
