+++ extensions/notion/src/utils/types.ts
@@ -1,78 +1,3 @@
-import type { Client } from "@notionhq/client";
-
-export interface User {
-  id: string;
-  type: string;
-  name: string | null;
-  avatar_url: string | null;
-}
-
-export interface Database {
-  id: string;
-  last_edited_time: number;
-  title: string | null;
-  icon_emoji: string | null;
-  icon_file: string | null;
-  icon_external: string | null;
-}
-
-// Currently supported properties
-const _supportedPropTypes = [
-  "title",
-  "number",
-  "rich_text",
-  "url",
-  "email",
-  "phone_number",
-  "date",
-  "checkbox",
-  "select",
-  "multi_select",
-  "formula",
-  "people",
-  "relation",
-  "status",
-] satisfies PagePropertyType["type"][];
-export const supportedPropTypes: PagePropertyType["type"][] = _supportedPropTypes;
-
-// all possible types:
-// "number" | "title" | "rich_text" | "url" | "email" | "phone_number" | "date" | "checkbox" | "select" | "formula" | "people" | "relation" | "multi_select" | "rollup" | "files" | "created_by" | "created_time" | "last_edited_by" | "last_edited_time"
-
-export interface DatabaseProperty {
-  id: string;
-  type: (typeof _supportedPropTypes)[number];
-  name: string;
-  options: DatabasePropertyOption[];
-  relation_id?: string;
-}
-
-export interface DatabasePropertyOption {
-  id?: string;
-  name: string;
-  color?: string;
-  icon?: string;
-}
-
-export interface Page {
-  object: "page" | "database";
-  id: string;
-  parent_page_id?: string;
-  parent_database_id?: string;
-  created_by?: string;
-  last_edited_time?: number;
-  last_edited_user?: string;
-  title: string | null;
-  icon_emoji: string | null;
-  icon_file: string | null;
-  icon_external: string | null;
-  url?: string;
-  properties: Record<string, PagePropertyType>;
-}
-
-export interface PageContent {
-  markdown: string | undefined;
-}
-
 export interface DatabaseView {
   properties?: Record<string, any>;
   create_properties?: string[];
@@ -94,8 +19,3 @@ export interface KabanView {
 export type UnwrapRecord<T> = T extends Record<never, infer U> ? U : never;
 export type UnwrapPromise<T> = T extends Promise<infer U> ? U : never;
 export type UnwrapArray<T> = T extends Array<infer U> ? U : never;
-
-export type NotionObject = UnwrapArray<UnwrapPromise<ReturnType<Client["search"]>>["results"]>;
-
-type NotionProperties<T, TObject> = T extends { object: TObject; properties: infer U } ? U : never;
-export type PagePropertyType = UnwrapRecord<NotionProperties<NotionObject, "page">>;
